requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery[redis]==5.5.3
redis==5.2.1
psycopg2-binary==2.9.10
openai==1.84.0
drf-spectacular==0.28.0
python-decouple==3.8
coverage==7.8.2
dj-database-url==2.3.0